+++ src/main/java/redstonedubstep/mods/vanishmod/mixin/MixinPlayerList.java
@@ -1,48 +1,58 @@
 package redstonedubstep.mods.vanishmod.mixin;
 
 import java.util.Arrays;
-import java.util.List;
 import java.util.UUID;
 
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import com.mojang.authlib.GameProfile;
+
 import net.minecraft.network.Connection;
 import net.minecraft.network.chat.ChatType;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientboundPlayerInfoPacket;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.server.network.ServerGamePacketListenerImpl;
 import net.minecraft.server.players.PlayerList;
 import net.minecraftforge.server.ServerLifecycleHooks;
 import redstonedubstep.mods.vanishmod.VanishUtil;
 
-@Mixin(PlayerList.class)
+@Mixin(value = PlayerList.class, priority = 900) //these mixin methods (especially the upper ones) are pretty important for this mod to work, so we set the priority of these mixins to be higher so no mixins clash because of same priorities
 public abstract class MixinPlayerList {
+	@Shadow
+	public abstract ServerPlayer getPlayer(UUID p_11260_);
+
 	@Unique
 	private ServerPlayer joiningPlayer;
 
-	//Remove vanished players from tab list information packet that is sent to all players on connect
-	@Redirect(method = "placeNewPlayer", at = @At(value = "NEW", target = "net/minecraft/network/protocol/game/ClientboundPlayerInfoPacket", ordinal = 0))
-	public ClientboundPlayerInfoPacket constructPacketToAll(ClientboundPlayerInfoPacket.Action actionIn, ServerPlayer[] playersIn) {
-		return new ClientboundPlayerInfoPacket(actionIn, VanishUtil.formatPlayerList(Arrays.asList(playersIn)));
+	//Prevent tab list information packet from being sent to other players if the joining player is vanished
+	@Redirect(method = "placeNewPlayer", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/players/PlayerList;broadcastAll(Lnet/minecraft/network/protocol/Packet;)V"))
+	public void constructAddPlayerPacketToAll(PlayerList instance, Packet<?> packet, Connection connection, ServerPlayer joiningPlayer) {
+		if (packet instanceof ClientboundPlayerInfoPacket infoPacket) {
+			GameProfile sentPlayer = !infoPacket.getEntries().isEmpty() ? infoPacket.getEntries().get(0).getProfile() : null;
+
+			if (sentPlayer == null || !VanishUtil.isVanished(instance.getPlayer(sentPlayer.getId())))
+				instance.broadcastAll(packet);
+		}
 	}
 
-	//Remove vanished players from tab list information packets that are sent to the joining player on connect. Includes an extra check to ensure that the vanished player gets information about itself
-	@Redirect(method = "placeNewPlayer", at = @At(value = "NEW", target = "net/minecraft/network/protocol/game/ClientboundPlayerInfoPacket", ordinal = 1))
-	public ClientboundPlayerInfoPacket constructPacketToJoinedPlayer(ClientboundPlayerInfoPacket.Action actionIn, ServerPlayer[] playersIn, Connection netManager, ServerPlayer receiver) {
-		List<ServerPlayer> list = Arrays.asList(playersIn);
+	//Prevent tab list information packets that contain vanished players from being sent to the joining player. Includes an extra check to ensure that a joining, vanished player gets information about itself
+	@Redirect(method = "placeNewPlayer", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/network/ServerGamePacketListenerImpl;send(Lnet/minecraft/network/protocol/Packet;)V", ordinal = 7))
+	public void constructPacketToJoinedPlayer(ServerGamePacketListenerImpl instance, Packet<?> packet, Connection connection, ServerPlayer joiningPlayer) {
+		if (packet instanceof ClientboundPlayerInfoPacket infoPacket) {
+			GameProfile sentPlayer = !infoPacket.getEntries().isEmpty() ? infoPacket.getEntries().get(0).getProfile() : null;
 
-		if (!VanishUtil.isVanished(receiver) || !receiver.equals(playersIn[0])) {
-			list = VanishUtil.formatPlayerList(list);
+			if (sentPlayer == null || !VanishUtil.isVanished(getPlayer(sentPlayer.getId())) || joiningPlayer.getUUID().equals(sentPlayer.getId()))
+				instance.send(packet);
 		}
-
-		return new ClientboundPlayerInfoPacket(actionIn, list);
 	}
 
 	//Remove vanished players from tab list information packet that is sent on disconnect